+++ es6/modules/expand-pair-trait.js
@@ -22,43 +22,128 @@ function getOpenCountChange(part) {
 	}
 }
 
-function getPairs(traits) {
+function match(start, end) {
+	return (
+		start != null &&
+		end != null &&
+		((start.part.location === "start" &&
+			end.part.location === "end" &&
+			start.part.value === end.part.value) ||
+			end.part.value === "")
+	);
+}
+
+function transformer(traits) {
+	let i = 0;
 	const errors = [];
-	let pairs = [];
-	if (traits.length === 0) {
-		return { pairs, errors };
-	}
-	let countOpen = 1;
-	const [firstTrait] = traits;
-	if (firstTrait.part.location === "start") {
-		for (let i = 1; i < traits.length; i++) {
-			const currentTrait = traits[i];
-			countOpen += getOpenCountChange(currentTrait.part);
-			if (countOpen === 0) {
-				const outer = getPairs(traits.slice(i + 1));
-				if (
-					currentTrait.part.value !== firstTrait.part.value &&
-					currentTrait.part.value !== ""
-				) {
-					errors.push(
-						getClosingTagNotMatchOpeningTag({
-							tags: [firstTrait.part, currentTrait.part],
-						})
-					);
-				} else {
-					pairs = [[firstTrait, currentTrait]];
-				}
+	while (i < traits.length) {
+		const part = traits[i].part;
+		if (part.location === "end") {
+			if (i === 0) {
+				traits.splice(0, 1);
+				errors.push(getUnmatchedLoopException(part));
 				return {
-					pairs: pairs.concat(outer.pairs),
-					errors: errors.concat(outer.errors),
+					traits,
+					errors,
 				};
 			}
+			const endIndex = i;
+			const startIndex = i - 1;
+			let offseter = 1;
+			if (match(traits[startIndex], traits[endIndex])) {
+				traits.splice(endIndex, 1);
+				traits.splice(startIndex, 1);
+				return { errors, traits };
+			}
+			while (offseter < 50) {
+				const startCandidate = traits[startIndex - offseter];
+				const endCandidate = traits[endIndex + offseter];
+				if (match(startCandidate, traits[endIndex])) {
+					traits.splice(endIndex, 1);
+					traits.splice(startIndex - offseter, 1);
+					return { errors, traits };
+				}
+				if (match(traits[startIndex], endCandidate)) {
+					traits.splice(endIndex + offseter, 1);
+					traits.splice(startIndex, 1);
+					return { errors, traits };
+				}
+				offseter++;
+			}
+			errors.push(
+				getClosingTagNotMatchOpeningTag({
+					tags: [traits[startIndex].part, traits[endIndex].part],
+				})
+			);
+			traits.splice(endIndex, 1);
+			traits.splice(startIndex, 1);
+			return {
+				traits,
+				errors,
+			};
 		}
+		if (traits[i] == null) {
+			break;
+		}
+		i++;
 	}
-	const { part } = firstTrait;
-	errors.push(getUnmatchedLoopException({ part, location: part.location }));
-	const outer = getPairs(traits.slice(1));
-	return { pairs: outer.pairs, errors: errors.concat(outer.errors) };
+	traits.forEach(function ({ part }) {
+		errors.push(getUnmatchedLoopException(part));
+	});
+	return {
+		traits: [],
+		errors,
+	};
+}
+
+function getPairs(traits) {
+	const levelTraits = {};
+	let errors = [];
+	let pairs = [];
+	let countOpen = 0;
+	let transformedTraits = [];
+
+	for (let i = 0; i < traits.length; i++) {
+		const currentTrait = traits[i];
+		const { part } = currentTrait;
+		const change = getOpenCountChange(currentTrait.part);
+		countOpen += change;
+		let level;
+		if (change === 1) {
+			level = countOpen - 1;
+		} else {
+			level = countOpen;
+		}
+		transformedTraits.push({ level, part });
+	}
+
+	while (transformedTraits.length > 0) {
+		const result = transformer(transformedTraits);
+		errors = errors.concat(result.errors);
+		transformedTraits = result.traits;
+	}
+
+	if (errors.length > 0) {
+		return { pairs, errors };
+	}
+	countOpen = 0;
+
+	for (let i = 0; i < traits.length; i++) {
+		const currentTrait = traits[i];
+		const { part } = currentTrait;
+		const change = getOpenCountChange(part);
+		countOpen += change;
+		if (change === 1) {
+			levelTraits[countOpen] = currentTrait;
+		} else {
+			const startTrait = levelTraits[countOpen + 1];
+			if (countOpen === 0) {
+				pairs = pairs.concat([[startTrait, currentTrait]]);
+			}
+		}
+		countOpen = countOpen >= 0 ? countOpen : 0;
+	}
+	return { pairs, errors };
 }
 
 const expandPairTrait = {
@@ -117,6 +202,9 @@ const expandPairTrait = {
 			lastPair = pair;
 			return [left, right];
 		});
+		if (errors.length > 0) {
+			return { postparsed, errors };
+		}
 
 		let currentPairIndex = 0;
 		let innerParts;